requirements.txt
asgiref==3.8.1
boto3==1.37.0
botocore==1.37.0
coverage==7.6.10
dj-database-url==2.3.0
Django==5.2.11
django-allauth==65.15.0
django-ckeditor==6.7.0
django-countries==7.6.0
django-crispy-forms==2.4
crispy-bootstrap5==0.7
django-jazzmin==4.1.0
django-js-asset==2.2.0
django-storages==1.14.6
django-ratelimit==4.1.0
gunicorn==23.0.0
jmespath==1.0.1
numpy==1.26.4
oauthlib==3.2.2
Pillow==12.1.1
psycopg2-binary==2.9.10
python3-openid==3.2.0
pytz==2025.2
requests-oauthlib==2.0.0
s3transfer==0.12.0
sqlparse==0.5.4
stripe==9.6.0
urllib3==2.6.3
pytest==8.3.4
pytest-django==4.9.0
